+++ tests/AvroConvertTests/Headless/HeadlessRecordsTests.cs
@@ -1,4 +1,5 @@
-﻿using FluentAssertions;
+﻿using System.Dynamic;
+using FluentAssertions;
 using SolTechnology.Avro;
 using Xunit;
 
@@ -8,7 +9,7 @@ namespace AvroConvertComponentTests.Headless
     public class HeadlessRecordsTests
     {
         [Fact]
-        public void Nullable_record_is_properly_Headless_Serialzied()
+        public void Anonymous_type_is_properly_headless_serialized()
         {
             //Arrange
             var valschema = @"
@@ -51,11 +52,58 @@ namespace AvroConvertComponentTests.Headless
             deserialized.Should().BeEquivalentTo(item);
 
         }
+
+        [Fact]
+        public void Expando_object_is_properly_headless_serialized()
+        {
+            //Arrange
+            var valschema = @"
+    {
+      ""type"": ""record"",
+      ""name"": ""TestObject"",
+      ""namespace"": ""ca.dataedu"",
+      ""fields"": [
+        {
+          ""name"": ""Value"",
+          ""type"":[
+            ""null"",
+            {
+              ""type"": ""record"",
+              ""name"": ""Value"",
+              ""fields"": [
+                {
+                  ""name"": ""KeyNested"",
+                  ""type"": ""string""
+                },
+                {
+                  ""name"": ""ValueNested"",
+                  ""type"": ""string""
+                }
+              ]
+            }]
+        }
+      ]
+    }
+";
+            dynamic item = new ExpandoObject();
+            item.value = new ExpandoObject();
+            item.value.KeyNested = "key1";
+            item.value.ValueNested = "val1";
+
+            //Act
+            byte[] value = AvroConvert.SerializeHeadless(item, valschema);
+            var deserialized = AvroConvert.DeserializeHeadless<AnonymousLikeClass?>(value, valschema);
+
+            //Assert
+            value.Should().NotBeNullOrEmpty();
+            deserialized.Should().BeEquivalentTo(new { Value = new { KeyNested = "key1", ValueNested = "val1" } });
+
+        }
     }
 
     public class AnonymousLikeClass
     {
-        public NestedClass Value { get; set; }
+        public NestedClass? Value { get; set; }
     }
 
     public class NestedClass